NN-4812: Restructure Routes dir
