+++ server/routes/adjudicationForReport/hearingDetails/index.ts
