+++ server/routes/adjudicationForReport/scheduleHearing/index.ts
