+++ server/routes/dis12Pdf/dis12Pdf.test.ts
@@ -6,7 +6,7 @@ import DecisionTreeService from '../../services/decisionTreeService'
 import PlaceOnReportService, { PrisonerResultSummary } from '../../services/placeOnReportService'
 import { AnswerType as Type } from '../../offenceCodeDecisions/Answer'
 import { OffenceRule } from '../../data/DraftAdjudicationResult'
-import AdjudicationPdf from './adjudicationPdf'
+import AdjudicationPdf from './dis12Pdf'
 import { ReportedAdjudication } from '../../data/ReportedAdjudicationResult'
 import { ConfirmedOnReportData } from '../../data/ConfirmedOnReportData'
 import { answer, question } from '../../offenceCodeDecisions/Decisions'
