+++ server/routes/index.ts
@@ -11,7 +11,7 @@ import prisonerRoutes from './prisonerRoutes'
 import prisonerSearchRoutes from './prisonerSearch'
 import prisonerSelectRoutes from './prisonerSelect'
 import yourCompletedReportsRoutes from './yourCompletedReports'
-import allCompletedReportsRoutes from './allCompletedReports'
+import allCompletedReportsRoutes from './viewAllHearingsAndReports/allReports'
 import continueReportSelectRoutes from './continueReport'
 
 import homepageRoutes from './homepage'
@@ -25,19 +25,19 @@ import detailsOfDamagesRoutes from './damages'
 import detailsOfEvidenceRoutes from './evidence'
 import detailsOfWitnessesRoutes from './witnesses'
 import associatedPrisonerRoutes from './associatedPrisoner'
-import hearingDetailsRoutes from './adjudicationTabbedParent/hearingDetails'
-import adjudicationReportRoutes from './adjudicationTabbedParent/prisonerReport'
-import scheduleHearingRoutes from './adjudicationTabbedParent/scheduleHearing'
-import viewScheduledHearingsRoutes from './viewScheduledHearings'
+import hearingDetailsRoutes from './adjudicationForReport/hearingDetails'
+import adjudicationReportRoutes from './adjudicationForReport/prisonerReport'
+import scheduleHearingRoutes from './adjudicationForReport/scheduleHearing'
+import viewScheduledHearingsRoutes from './viewAllHearingsAndReports/allHearings'
 import acceptedReportConfirmationRoutes from './acceptedReportConfirmation'
 import selectGenderRoutes from './selectGender'
 import confirmDISFormsIssuedRoutes from './confirmDISFormsIssued'
 import addDateAndTimeOfIssueRoutes from './addDateAndTimeOfIssue'
 
 import { Services } from '../services'
-import adjudicationPdfRoutes from './adjudicationPdf'
+import adjudicationPdfRoutes from './dis12Pdf'
 import adjudicationUrls from '../utils/urlGenerator'
-import printCompletedDISFormsRoutes from './printCompletedDISForms'
+import printCompletedDISFormsRoutes from './printCompletedDIS12Forms'
 
 export default function routes(
   router: Router,
