+++ server/routes/viewAllHearingsAndReports/allHearings/index.ts
@@ -1,11 +1,11 @@
 import express, { RequestHandler, Router } from 'express'
-import asyncMiddleware from '../../middleware/asyncMiddleware'
+import asyncMiddleware from '../../../middleware/asyncMiddleware'
 
 import ScheduledHearings from './viewScheduledHearings'
 
-import ReportedAdjudicationsService from '../../services/reportedAdjudicationsService'
-import UserService from '../../services/userService'
-import adjudicationUrls from '../../utils/urlGenerator'
+import ReportedAdjudicationsService from '../../../services/reportedAdjudicationsService'
+import UserService from '../../../services/userService'
+import adjudicationUrls from '../../../utils/urlGenerator'
 
 export default function viewScheduledHearingsRoutes({
   reportedAdjudicationsService,
