+++ server/routes/viewAllHearingsAndReports/allReports/allCompletedReports.ts
@@ -1,11 +1,11 @@
 import { Request, Response } from 'express'
-import mojPaginationFromPageResponse, { pageRequestFrom } from '../../utils/mojPagination/pagination'
-import { hasAnyRole } from '../../utils/utils'
-import { ReportedAdjudicationEnhanced } from '../../data/ReportedAdjudicationResult'
-import { ApiPageResponse } from '../../data/ApiData'
-import ReportedAdjudicationsService from '../../services/reportedAdjudicationsService'
-import UserService from '../../services/userService'
-import adjudicationUrls from '../../utils/urlGenerator'
+import mojPaginationFromPageResponse, { pageRequestFrom } from '../../../utils/mojPagination/pagination'
+import { hasAnyRole } from '../../../utils/utils'
+import { ReportedAdjudicationEnhanced } from '../../../data/ReportedAdjudicationResult'
+import { ApiPageResponse } from '../../../data/ApiData'
+import ReportedAdjudicationsService from '../../../services/reportedAdjudicationsService'
+import UserService from '../../../services/userService'
+import adjudicationUrls from '../../../utils/urlGenerator'
 import {
   fillInDefaults,
   filterFromUiFilter,
@@ -14,8 +14,8 @@ import {
   uiFilterFromBody,
   uiFilterFromRequest,
   validate,
-} from '../../utils/adjudicationFilterHelper'
-import { FormError } from '../../@types/template'
+} from '../../../utils/adjudicationFilterHelper'
+import { FormError } from '../../../@types/template'
 
 export default class AllCompletedReportsRoutes {
   constructor(
