+++ server/services/reportedAdjudicationsService.ts
@@ -37,7 +37,7 @@ import {
   HearingDetails,
 } from '../data/DraftAdjudicationResult'
 import LocationService from './locationService'
-import { ReviewStatus } from '../routes/adjudicationTabbedParent/prisonerReport/prisonerReportReviewValidation'
+import { ReviewStatus } from '../routes/adjudicationForReport/prisonerReport/prisonerReportReviewValidation'
 import { PrisonerResultSummary } from './placeOnReportService'
 import PrisonerSimpleResult from '../data/prisonerSimpleResult'
 import { Alert, alertFlagLabels, AlertFlags } from '../utils/alertHelper'
